vue3ts-pywebview/requirements.txt
pywebview==6.1
pyinstaller==6.14.1
nuitka==2.7.12